backend/pyproject.toml
[project]
name = "mindquest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-restful>=0.3.10",
    "flask-security-too>=5.6.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-whooshee>=0.9.1",
    "requests>=2.32.4",
]
